requirements.txt
h5py==3.9.0
networkx==2.8.7
networkx==2.8.4
numpy==1.24.4
numpy==1.23.5
python_graph_core==1.8.2
python_igraph==0.11.6
scikit_image==0.19.3
scipy==1.9.1
